refactor(backend): centralize experience and education fixtures in conftest.py
