+++ backend/tests/routers/admin/test_education.py
@@ -1,30 +1,9 @@
-import pytest
 from fastapi.testclient import TestClient
 from sqlmodel import Session, select
 
-from app.models import Education, Profile
+from app.models import Education
 
 ADMIN_EDUCATION_URL = "/api/admin/education"
-
-
-@pytest.fixture
-def education(session: Session) -> Education:
-    profile = Profile()
-    session.add(profile)
-    session.commit()
-    session.refresh(profile)
-
-    education = Education(
-        school="MIT",
-        degree="Bachelor",
-        year=2022,
-        profile_id=profile.id,
-    )
-    session.add(education)
-    session.commit()
-    session.refresh(education)
-
-    return education
 
 
 def test_create_education(admin_client: TestClient, session: Session) -> None:
