+++ backend/tests/routers/admin/test_experiences.py
@@ -1,32 +1,9 @@
-from datetime import date
-
-import pytest
 from fastapi.testclient import TestClient
 from sqlmodel import Session, select
 
-from app.models import Experience, Profile
+from app.models import Experience
 
 ADMIN_EXPERIENCE_URL = "/api/admin/experiences"
-
-
-@pytest.fixture
-def experience(session: Session) -> Experience:
-    profile = Profile()
-    session.add(profile)
-    session.commit()
-    session.refresh(profile)
-
-    experience = Experience(
-        company="Acme",
-        position="backend developer",
-        start_date=date(2022, 1, 1),
-        profile_id=profile.id,
-    )
-    session.add(experience)
-    session.commit()
-    session.refresh(experience)
-
-    return experience
 
 
 def test_create_experience(admin_client: TestClient, session: Session) -> None:
